Store context in a thread-safe variable.
Currently, each Endpoint instance carries the pale context
in an instance variable called `_context`.

Because the Endpoint instance is global, and not thread-local,
we need to store the `_context` object in a thread-safe variable
provided by threading.local().

Additionally, a `_finally()` method is added to `Endpoint` and
is called when `_execute()` completes, regardless of success.
This is intended for any cleanup the Endpoint may need to do.

This patch also provides docstring updates to various methods.